requirements.txt
requests
beautifulsoup4
ipwhois
dnspython
python-whois
